requirements.txt
google-generativeai
streamlit
arxiv
python-dotenv
pypdf
chromadb
